services/backend/requirements.txt
fastapi
uvicorn
tortoise-orm
aerich
asyncpg
databases[asyncio]
psycopg2-binary
SQLAlchemy==2.0.15

# dev
pytest
httpx
